requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LINE SDK (latest version)
line-bot-sdk>=3.11.0

# HTTP Client
httpx>=0.26.0

# RSS Parser
feedparser>=6.0.10

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0

# Scheduler
apscheduler>=3.10.4

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Timezone
pytz>=2024.1
